Add the ability to deploy kube master services as containerized
Add logic to allow deploying kubernetes using the built-in RPMS or as
containers.  Additionally, add the ability to skip the AH upgrade
process.

The script defaults to upgrading AH to the latest ostree verion, and
to deploy containerized kubernetes master services (apiserver,
controller-manager, and scheduler).  To override these, the following
variables can be set in the environment:

* CONTAINERIZED=false # Will deploy the normal RPM-based kubernetes master services
* AHUPGRADE=false     # Will skip the "atomic host upgrade" step